moving comparison package to core
