+++ src/main/java/fr/inria/coming/core/filter/diff/syntcomparison/Fragmentable.java
@@ -1,4 +1,4 @@
-package comparison;
+package fr.inria.coming.core.filter.diff.syntcomparison;
 
 import org.eclipse.compare.rangedifferencer.IRangeComparator;
