+++ src/main/java/fr/inria/coming/core/filter/diff/syntcomparison/SJavaTokenComparator.java
@@ -1,4 +1,4 @@
-package comparison;
+package fr.inria.coming.core.filter.diff.syntcomparison;
 
 /*******************************************************************************
  * Copyright (c) 2000, 2008 IBM Corporation and others.
