+++ src/main/java/fr/inria/coming/core/filter/files/CommitSizeFilter.java
@@ -10,6 +10,7 @@ import fr.inria.coming.core.interfaces.IFilter;
 
 public class CommitSizeFilter extends AbstractChainedFilter {
 
+	
 	public CommitSizeFilter() {
 		super();
